fix: move delete, update method to bookmarkCommandService file
