+++ smeem-application/src/main/java/com/smeem/application/domain/bookmark/BookmarkQueryService.java
@@ -4,11 +4,7 @@ import com.smeem.application.domain.bookmark.model.Bookmark;
 import com.smeem.application.port.input.bookmark.BookmarkQueryUseCase;
 import com.smeem.application.port.input.bookmark.dto.BookmarkDetailResponse;
 import com.smeem.application.port.input.bookmark.dto.BookmarkListResponse;
-import com.smeem.application.port.input.bookmark.dto.BookmarkModifyResponse;
-import com.smeem.application.port.input.bookmark.dto.BookmarkUpdateRequest;
 import com.smeem.application.port.output.persistence.BookmarkPort;
-import com.smeem.common.exception.ExceptionCode;
-import com.smeem.common.exception.SmeemException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BookmarkQueryService implements BookmarkQueryUseCase {
 
@@ -37,26 +33,5 @@ public class BookmarkQueryService implements BookmarkQueryUseCase {
                 .sorted((a, b) -> b.getCreatedAt().compareTo(a.getCreatedAt()))
                 .toList();
         return BookmarkListResponse.from(bookmarks);
-    }
-
-    @Override
-    public void deleteBookmark(long memberId, long bookmarkId) {
-        Bookmark bookmark = bookmarkPort.getById(bookmarkId);
-        validateBookmarkOwnership(bookmark.getMemberId(), memberId);
-        bookmarkPort.deleteByBookmarkId(bookmarkId);
-    }
-
-    @Override
-    public BookmarkModifyResponse updateBookmark(long memberId, long bookmarkId, BookmarkUpdateRequest request) {
-        Bookmark bookmark = bookmarkPort.getById(bookmarkId);
-        validateBookmarkOwnership(bookmark.getMemberId(), memberId);
-        Bookmark updatedBookmark = bookmarkPort.update(request.toDomain(bookmark));
-        return BookmarkModifyResponse.from(updatedBookmark);
-    }
-
-    private void validateBookmarkOwnership(long bookmarkWriterId, long memberId) {
-        if (bookmarkWriterId != memberId) {
-            throw new SmeemException(ExceptionCode.INVALID_MEMBER_AND_BOOKMARK);
-        }
     }
 }
